Moved scene-specific selector into service directory
